// src/app/interfaces/prodist.ts
export interface PRODIST {
    ContaDeLuz: string;
    NumeroDoCliente: String;
    ClasseUC: String;
    CotaMensal: String;
    PotenciaInstalada: String;
    TensaoDeAtendimento: String;
    TipoDeConexao: String;
    TipoDeRamal: String;
    PotenciaInstaladaGeral: String;
    TipoDaFonteDeGeracao: String;
    MenorConsumoUltimos12: String;
    MaiorConsumoUltimos12: String;
}